pyproject.toml
[project]
name = "discord-slack-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Yan",email = "119687802+sndmndss@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "slack-sdk (>=3.35.0,<4.0.0)",
    "slack-bolt (>=1.23.0,<2.0.0)",
    "peewee (>=3.18.2,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
